+++ pyespn/classes/image.py
@@ -1,5 +1,8 @@
+from pyespn.core.decorators import validate_json
+import requests
 
 
+@validate_json("image_json")
 class Image:
     """
     Represents an image object from the ESPN API, typically associated with players, teams, or events.
@@ -14,6 +17,19 @@ class Image:
         alt (str): Alternative text for the image.
         rel (list): A list of roles describing the image (e.g., "default", "profile").
         last_updated (str): The last updated timestamp of the image.
+
+    Methods:
+        __init__(image_json, espn_instance):
+            Initializes the Image object using JSON data and a reference to the ESPN API wrapper.
+
+        __repr__():
+            Returns a string representation of the Image object.
+
+        _load_image_data():
+            Parses and loads image metadata from the provided JSON data.
+
+        load_image() -> bytes:
+            Downloads and returns the binary content of the image from the reference URL.
     """
 
     def __init__(self, image_json, espn_instance):
@@ -29,9 +45,12 @@ class Image:
         self.espn_instance = espn_instance
         self._load_image_data()
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         """
         Returns a string representation of the Image instance.
+
+        Returns:
+            str: A string showing the image's name.
         """
         return f"<Image | {self.name}>"
 
@@ -48,3 +67,23 @@ class Image:
             self.name = self.alt
         self.rel = self.image_json.get('rel')
         self.last_updated = self.image_json.get('lastUpdated')
+
+    def load_image(self) -> bytes:
+        """
+        Downloads and returns the image content from the object's reference URL.
+
+        Returns:
+            bytes: The binary content of the image.
+        """
+        image_request = requests.get(self.ref)
+        image = image_request.content
+        return image
+
+    def to_dict(self) -> dict:
+        """
+        Converts the Image instance to its original JSON dictionary.
+
+        Returns:
+            dict: The images's raw JSON data.
+        """
+        return self.image_json